+++ src/types/config.d.ts
@@ -1,0 +1,7 @@
+type ExcludeFolders = string[];
+type FolderTermPaths = { [folder: string]: string[] };
+
+interface Config {
+  exclude_folders: ExcludeFolders;
+  folder_term_paths: FolderTermPaths;
+}